Resolved not understanding versions of noun in Hindi
Wrote an implementation of Hamming Distance (https://en.wikipedia.org/wiki/Hamming_distance) to resolve the issue of flagging 'kela' and 'kele' different. This will, however, flag 'kela' and 'mela' the same which is highly problematic at this point. Will have to figure out a way to resolve this bug.