Handled multiple code-breaks when testing different languages.
The object recognition for Spanish was breaking because of the way 'the' inflects based on the gender of the object. The code now accounts for that by ignoring two-letter words that change when checking for objects. Accordingly, the noun that was taken from the object also accommodates for this fact and takes the appropriate word for comparison before looking for the preposition. Fixed a huge bug (actually small) in the way adjective-noun positioning was decided after faulty outputs for Bengali and Odia. Have to make language arrays from beta-participant inputs.