Of multiple changes and lost code.
Handling errors better now. When Gina can't figure out SVO word order due to faulty input, she says so now (and failure to get the subject fails adjectives too. Will change it later). Midway through the update, I found that the Kannada update from yesterday was missing. Added the Kannada fix (noun inflection is beyond the Jaro-Winkler threshold which causes preposition to fail since it can't understand that the noun is the same but with morphemes. The fix is to just check what remains the same in sentence 1 and 3 and does not use the noun from sentence 0. Whatever changes is the preposition and its relative positioning with what does not change (noun) gives us the ordering. Will do something similar for adjectives in a while.